refactor: interface 선언 분리 및 리팩토링
