+++ frontend/src/components/Header/SearchBar/SearchBar.style.ts
@@ -1,6 +1,8 @@
 import { Container } from "@mui/material";
 import { styled } from "@mui/material/styles";
 
+import { SearchBarContainerProps } from "@types";
+
 const SearchBarContainer = styled(Container, {
   shouldForwardProp: (prop) => prop !== "currentPage",
   name: "MyThemeComponent",
@@ -18,7 +20,3 @@ const SearchBarContainer = styled(Container, {
 );
 
 export default SearchBarContainer;
-
-interface SearchBarContainerProps {
-  currentPage?: string;
-}
