+++ frontend/src/components/Header/SearchBar/SelectItemArea/SelectItemTemplate/SelectItemTemplate.style.ts
@@ -1,14 +1,12 @@
-import { Grid, GridProps } from "@mui/material";
+import { Grid } from "@mui/material";
 import { styled } from "@mui/material/styles";
 
-const GridItem = styled(Grid)<GridItemProps>(
+import { SelectItemTemplateProps } from "@types";
+
+const GridItem = styled(Grid)<SelectItemTemplateProps>(
   ({ divide, theme: { palette } }) => `
   border-right: ${divide && `1px solid ${palette.grey5.main}`}};
 `
 );
 
 export default GridItem;
-
-interface GridItemProps extends GridProps {
-  divide?: string;
-}
